pyproject.toml
[tool.poetry]
name = "protein-rebuilder"
version = "0.1.0"
description = ""
authors = ["Jules <161369871+google-labs-jules[bot]@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "protein_rebuilder"}, {include = "scripts"}]

[tool.poetry.dependencies]
python = ">=3.12"
biopython = ">=1.85,<2.0"
numpy = ">=2.3.2,<3.0.0"
pdbfixer = {git = "https://github.com/openmm/pdbfixer.git"}
openmm = ">=8.3.1,<9.0.0"
pandas = "^2.3.2"

[tool.poetry.scripts]
setup-dunbrack = "scripts.fetch_dunbrack:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.12"
pylint = "^3.3.8"
mypy = "^1.17.1"
pytest = "^8.4.2"
deptry = "^0.17.0"

[tool.mypy]
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["openmm", "pdbfixer"]
ignore_missing_imports = true
